# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Pawnshop Management System API",
    "description": "A comprehensive API for managing pawnshop operations including loans, customers, items, payments, and sales.",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "email": "support@pawnshop.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8090",
  "basePath": "/api/v1",
  "schemes": ["http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Type 'Bearer' followed by a space and the JWT token."
    }
  },
...
  "paths": {
...
    "/items": {
      "get": {
        "tags": ["Items"],
        "summary": "List Items",
        "description": "Get a paginated list of items",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "page", "in": "query", "type": "integer", "default": 1},
          {"name": "per_page", "in": "query", "type": "integer", "default": 20},
          {"name": "branch_id", "in": "query", "type": "integer"},
          {"name": "status", "in": "query", "type": "string"},
          {"name": "search", "in": "query", "type": "string"}
        ],
        "responses": {
          "200": {
            "description": "List of items"
          }
        }
      },
      "post": {
        "tags": ["Items"],
        "summary": "Create Item",
        "description": "Create a new item",
        "security": [{"BearerAuth": []}],
        "responses": {
          "201": {
            "description": "Item created"
          }
        }
      }
    },
...
    "/branches": {
      "get": {
        "tags": ["Branches"],
        "summary": "List Branches",
        "description": "Get a list of all branches",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of branches"
          }
        }
      },
      "post": {
        "tags": ["Branches"],
        "summary": "Create Branch",
        "description": "Create a new branch",
        "security": [{"BearerAuth": []}],
        "responses": {
          "201": {
            "description": "Branch created"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "List Categories",
        "description": "Get a list of item categories",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of categories"
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Create Category",
        "description": "Create a new category",
        "security": [{"BearerAuth": []}],
        "responses": {
          "201": {
            "description": "Category created"
          }
        }
      }
    },
...
  },
...
}